Rangelinien sind nun richtig
git-svn-id: svn+ssh://elmi.iti.uni-luebeck.de/export/svn/Hanse/branches/threadv1@563 8f13a6d9-6203-4c43-9b9f-a3039379152f

+++ Hanse/Module_EchoSounder/echosounderform.cpp
@@ -71,7 +71,7 @@ void EchoSounderForm::updateSounderView(const EchoReturnData data)
     float n = data.getEchoData().length();
     scene.clear();
     int height = ui->graphicsView->height()-1;
-    int faktor = (height / range);
+    float faktor = (height / range);
     if(range >= 20.0){
         faktor = faktor * 10;
     }
